Bug counting dataset on view groups
## Description
If we add a dataset to 2 groups or more to the field EVA (using og_count) show a incorrect value. For example:
We have 2 groups and we add a dataset linked with both groups. Then on the section groups we will see 2 dataset instead 1.


## Steps to Reproduce

Create a dataset and associate with 2 groups. On the groups section the total dataset count is a bad value.

## Acceptance Criteria

Create a dataset and associate with 2 groups. On the groups section the total datasets should be fine.