Update dkan_sitewide to set the default admin_theme to default
As a site maintainer I do not want the dkan_sitewide feature to show as overridden when using a custom sub-theme for the admin theme. This value should be set 'Default theme' (0) as this works for both NuBoot Radix and any subtheme. No reason to have it hard coded as nuboot_radix.