Use correct term table fields. (#1486)
REF https://jira.govdelivery.com/browse/CIVIC-4703

Description
===
The current behat scenario "Scenario: See topics on the homepage as anonymous
user" is testing a mute test becuase it checks for an icon that is already added
by the dkan profile on installation.  The problem is that on older sites that do
not run the profile, the default icon is not added revealing a test bug when
this icon is checked for.  Although the test adds the icon a background step, it
fails to correctly add the icon font and icon type because the table field are
not correctly mapped or the correct table field name is not used.

This commit fixes two issues.
- [ ] Make sure to use an icon in the test that is not already added to the
  front page, that way if there is an uderlying issue it is caught but the
  upstream tests.

- [ ] Make sure to use the correct field labels for the background table.