Allow dkanextension to handle custom fields.
Description
===
Currently behat tests throw errors when fields are required and not accounted
for in the standard tests. We need to be able to handle those cases so that dkan
tests work under custom conditions. We should create a ticket for unraveling
custom tests that we created for clients (HHS) once we have this solution in
place.

AC
==
- [ ] when creating a node in a feature, any field that is required (custom or
    non-custom) will automatically be populated and accounted for in the test