Allow dkanextension to handle custom fields. (#1963)
* Allow dkanextension to handle custom fields.

Description
===
Currently behat tests throw errors when fields are required and not accounted
for in the standard tests. We need to be able to handle those cases so that dkan
tests work under custom conditions. We should create a ticket for unraveling
custom tests that we created for clients (HHS) once we have this solution in
place.

AC
==
- [ ] when creating a node in a feature, any field that is required (custom or
    non-custom) will automatically be populated and accounted for in the test

* Fix devel_generate does not handle author field.

If author field generated it must be part of table or skipped.

* Auto add required references to dataset rest_api table.

REF CIVIC-6519

* Fix set_field throws when field type "list<text>":

dkanextension currently throws when setting a field of type "list<text>".  This
changes fixes that issue.

* Fix test fail when description auto-generated.

* Update DatasetContext.php

https://github.com/NuCivic/dkan_starter/issues/332

* Update DatasetContext.php

* Add documentation for changes made regarding required fields.

* Update changelog

* Skip linting on dkanextension files

Update dkanextension readme file