Add behat step method to handle custom inputs.
REF CIVIC-6639

This change adds a new step method

"When I fill-in with:"

Which takes a table with field and values and inserts the values to specified
field in a dataset node add form with the additional feature of also adding any
missing required inputs.

By using this new step we guard against test failure in the node add context
when an implementation either adds a new custom required field or makes a field
required that is not accounted for in the test.

QA Steps
==
- [ ]  admin/dkan/dataset_forms and enable "Validate dataset form according to
  Project Open Data"
- [ ]  Locally run the markdown_editor.feature tests, they should continue to
  pass.